go back to text/template, s/draft/autoidx/gi
html/template is too protective to be useful for this application,
so I'm moving back to text/template and littering files with the
{{ . |html }} functions to make sure things are escaped in critical
areas.

Renamed the draft flag and related code to autoIdx, so it has a simpler
meaning now. autoidx: true (the default) means "include this page in
the list of pages used for index.html and rss.xml (and similar)".
Setting it to false will get the page ignored, which can be used for
working on drafts and for pages that I never want to be in that index.

There are still some issues with the imported html from blogger, but
that is probably best fixed by running all of that content through
html tidy, which will also unwrap the long single lines of html.